+++ src/main/java/com/induamerica/backend/repository/CargaRepository.java
@@ -3,6 +3,6 @@ package com.induamerica.backend.repository;
 import com.induamerica.backend.model.Carga;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CargaRepository extends JpaRepository<Carga, Integer> {
+public interface CargaRepository extends JpaRepository<Carga, Long> {
     boolean existsByCodigoCarga(String codigoCarga);
 }
