+++ src/main/java/com/induamerica/backend/repository/LocalRepository.java
@@ -3,5 +3,5 @@ package com.induamerica.backend.repository;
 import com.induamerica.backend.model.Local;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocalRepository extends JpaRepository<Local, Integer> {
+public interface LocalRepository extends JpaRepository<Local, Long> {
 }
